Analyse parquet
