+++ STM_Services/STM_Analyse_Daily_Stops_Data/requirements.txt
@@ -1,5 +1,5 @@
-polars==0.19.19
-pandas==2.1.3
+boto3==1.28.77
 fastparquet==2023.10.1
+pandas==2.1.4
+polars==0.19.19
 pytz==2023.3.post1
-numpy==1.26.0
